add web module
